refactor: split NavTab from CourseTabsNavigation
